fix(studio): fix lint, split BandRenderer, use testids in RangeBand tests
- Split BandRenderer into its own file so each file has a single export type,
  satisfying react-refresh/only-export-components without eslint-disable
- Add data-testid to path and clipPath elements; replace container.querySelector
  with screen.queryByTestId/getByTestId in tests
- Barrel index.tsx now only re-exports the hook (BandRenderer is an impl detail)
- Fix import order in useRangeBand.tsx (@studio sorts before react alphabetically)
- Restore nvm use 22 in post-edit hook (was dropped during rebase)
