+++ web/packages/studio/src/components/charts/RangeBand/index.tsx
@@ -1,5 +1,5 @@
 // SPDX-FileCopyrightText: Copyright (c) 2025-2026 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
 // SPDX-License-Identifier: Apache-2.0
 
-export { BandRenderer, useRangeBand } from '@studio/components/charts/RangeBand/useRangeBand';
-export type { BandRendererProps, UseRangeBandOptions } from '@studio/components/charts/RangeBand/useRangeBand';
+export { useRangeBand } from '@studio/components/charts/RangeBand/useRangeBand';
+export type { UseRangeBandOptions } from '@studio/components/charts/RangeBand/useRangeBand';
